src/requirements.txt
python-telegram-bot
APScheduler
python-dotenv
telegram